added password modification

+++ src/controllers/customersController.js
@@ -25,7 +25,7 @@ export const createClient = async (req, res) => {
 export const getAllClients = async (req, res) => {
   try {
     const clients = await prisma.customers.findMany({
-      orderBy: { id: 'asc' }
+      orderBy: { id: 'desc' }
     });
     res.status(200).json(clients);
   } catch (error) {
@@ -70,19 +70,32 @@ export const updateClient = async (req, res) => {
 };
 
 export const deleteClient = async (req, res) => {
- const { id } = req.params;
+  const { id } = req.params;
 
   try {
+    // Vérifiez si le client a des commandes liées
     const linkedOrders = await prisma.orders.findMany({
       where: { customerId: parseInt(id) },
     });
 
     if (linkedOrders.length > 0) {
       return res.status(400).json({
-        message: ('Unable to delete a customer due to a sale or reception'),
+        message: i18next.t('customer.hasOrders'),
       });
     }
 
+    // Vérifiez si le client a des ventes liées
+    const linkedSales = await prisma.sales.findMany({
+      where: { customerId: parseInt(id) },
+    });
+
+    if (linkedSales.length > 0) {
+      return res.status(400).json({
+        message: i18next.t('customer.hasSales'),
+      });
+    }
+
+    // Si aucune relation n'est trouvée, supprimez le client
     await prisma.customers.delete({
       where: { id: parseInt(id) },
     });
